+++ packages/router/src/operators/apply_redirects.ts
@@ -11,7 +11,7 @@ import {MonoTypeOperatorFunction} from 'rxjs';
 import {map, switchMap} from 'rxjs/operators';
 
 import {applyRedirects as applyRedirectsFn} from '../apply_redirects';
-import {Routes} from '../config';
+import {Routes} from '../models';
 import {NavigationTransition} from '../router';
 import {RouterConfigLoader} from '../router_config_loader';
 import {UrlSerializer} from '../url_tree';
