+++ packages/router/src/shared.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {Route, UrlMatchResult} from './config';
+import {Route, UrlMatchResult} from './models';
 import {UrlSegment, UrlSegmentGroup} from './url_tree';
