+++ packages/router/test/helpers.ts
@@ -8,7 +8,7 @@
 
 import {Type} from '@angular/core';
 
-import {Data, ResolveData, Route} from '../src/config';
+import {Data, ResolveData, Route} from '../src/models';
 import {ActivatedRouteSnapshot} from '../src/router_state';
 import {convertToParamMap, ParamMap, Params, PRIMARY_OUTLET} from '../src/shared';
 import {equalSegments, UrlSegment, UrlSegmentGroup, UrlTree} from '../src/url_tree';
