+++ packages/router/test/router_preloader.spec.ts
@@ -13,7 +13,7 @@ import {BehaviorSubject, Observable, of, throwError} from 'rxjs';
 import {catchError, delay, filter, switchMap, take} from 'rxjs/operators';
 
 import {Route, RouteConfigLoadEnd, RouteConfigLoadStart, Router, RouterModule} from '../index';
-import {LoadedRouterConfig} from '../src/config';
+import {LoadedRouterConfig} from '../src/models';
 import {RouterTestingModule} from '../testing';
